fix : finish Partner Page fix merge error #12

+++ src/main/java/com/sinse/universe/dto/response/PromotionResponse.java
@@ -9,7 +9,7 @@ public record PromotionResponse(
         String name,
         String description,
         String img,
-        BigDecimal price,
+        Integer price,
         boolean fanOnly,
         int stockQty,
         int limitPerUser,
